backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
numpy
requests
pydantic
typing-extensions
